Fix build errors and warnings across platforms
Project-owned code:
- vm_x86.c: Add trailing comma to DROP() calls for -Wgnu-zero-variadic-macro-arguments
- sdl_icon.h: Add newline at end of file
- ui_css.c: Remove unused variable 'a'
- cl_cin_ffmpeg.c: Fix swr_convert pointer qualifier (pragma for safe cast)
- tr_main.c: Fix sign-compare (unsigned loop vars), cast-qual (const sortKey)
- tr_mesh.c: Use unsigned int for num_dlights loop
- tr_model.c: Cast sizeof for fileSize comparison
- tr_model_iqm.c: Fix IQM_CheckRange sign/const comparisons

External libs (CMake suppression):
- FLAC: -Wno-cast-qual -Wno-format-nonliteral -Wno-static-in-inline
- zlib: -Wno-format-nonliteral
- Duktape: -Wno-format-nonliteral -Wno-unused-but-set-variable
- Opus: -Wno-cast-align

+++ src/renderers/opengl/tr_model.c
@@ -71,7 +71,7 @@ static qhandle_t R_RegisterMD3(const char *name, model_t *mod)
 		if ( !buf.v )
 			continue;
 
-		if ( fileSize < sizeof( md3Header_t ) ) {
+		if ( fileSize < (int)sizeof( md3Header_t ) ) {
 			ri.Printf( PRINT_WARNING, "%s: truncated header for %s\n", __func__, name );
 			ri.FS_FreeFile( buf.v );
 			break;
